+++ src/editor/editor-sprites.tsx
