+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react"
 import { createRoot } from "react-dom/client"
 import { App } from "./app"
-import { EditorStore } from "./editor-store"
+import { EditorStore } from "./editor/editor-store"
 import "./tailwind.css"
 
 createRoot(document.querySelector("#root")!).render(
